Removing bugs and adding delete operation
