+++ Fotosopic/Fotosopic.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 4.8.0, 2018-12-30T23:17:26. -->
+<!-- Written by QtCreator 4.8.0, 2019-01-01T23:23:31. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
