+++ Fotosopic/forms/mainwindow.ui
@@ -76,7 +76,7 @@
   <widget class="QToolBar" name="mainToolBar">
    <property name="minimumSize">
     <size>
-     <width>50</width>
+     <width>0</width>
      <height>0</height>
     </size>
    </property>
@@ -102,7 +102,7 @@
     <enum>LeftToolBarArea</enum>
    </attribute>
    <attribute name="toolBarBreak">
-    <bool>true</bool>
+    <bool>false</bool>
    </attribute>
    <addaction name="action_Zoom"/>
    <addaction name="action_Zoom_out"/>
@@ -118,8 +118,8 @@
   <widget class="QStatusBar" name="statusBar"/>
   <action name="action_Open">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/open.png</normaloff>:/images/open.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/open.png</normaloff>:/icons/icons/open.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Open</string>
@@ -127,8 +127,8 @@
   </action>
   <action name="action_Save">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/save.png</normaloff>:/images/save.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/save.png</normaloff>:/icons/icons/save.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Save</string>
@@ -136,8 +136,8 @@
   </action>
   <action name="action_Save_as">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/save.png</normaloff>:/images/save.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/save.png</normaloff>:/icons/icons/save.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Save as...</string>
@@ -145,8 +145,8 @@
   </action>
   <action name="action_Undo">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/undo.png</normaloff>:/images/undo.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/undo.png</normaloff>:/icons/icons/undo.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Undo</string>
@@ -154,8 +154,8 @@
   </action>
   <action name="action_Redo">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/redo.png</normaloff>:/images/redo.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/redo.png</normaloff>:/icons/icons/redo.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Redo</string>
@@ -168,8 +168,8 @@
   </action>
   <action name="action_Delete">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/delete.png</normaloff>:/images/delete.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/delete.png</normaloff>:/icons/icons/delete.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Delete</string>
@@ -177,8 +177,8 @@
   </action>
   <action name="action_Exit">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/exit.png</normaloff>:/images/exit.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/exit.png</normaloff>:/icons/icons/exit.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Exit</string>
@@ -186,8 +186,8 @@
   </action>
   <action name="action_Zoom">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/zoom.png</normaloff>:/images/zoom.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/zoom.png</normaloff>:/icons/icons/zoom.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Zoom</string>
@@ -195,8 +195,8 @@
   </action>
   <action name="action_Zoom_out">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/zoom_out.png</normaloff>:/images/zoom_out.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/zoom_out.png</normaloff>:/icons/icons/zoom_out.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Zoom out</string>
@@ -204,8 +204,8 @@
   </action>
   <action name="action_Resize">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/resize.png</normaloff>:/images/resize.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/resize.png</normaloff>:/icons/icons/resize.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Resize</string>
@@ -213,8 +213,8 @@
   </action>
   <action name="action_Mirror">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/mirror.png</normaloff>:/images/mirror.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/mirror.png</normaloff>:/icons/icons/mirror.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Mirror</string>
@@ -222,8 +222,8 @@
   </action>
   <action name="action_Draw">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/draw.png</normaloff>:/images/draw.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/draw.png</normaloff>:/icons/icons/draw.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Draw</string>
@@ -231,8 +231,8 @@
   </action>
   <action name="action_Text">
    <property name="icon">
-    <iconset resource="resources.qrc">
-     <normaloff>:/images/text.png</normaloff>:/images/text.png</iconset>
+    <iconset resource="../resources/icons.qrc">
+     <normaloff>:/icons/icons/text.png</normaloff>:/icons/icons/text.png</iconset>
    </property>
    <property name="text">
     <string>&amp;Text</string>
@@ -241,7 +241,7 @@
  </widget>
  <layoutdefault spacing="6" margin="11"/>
  <resources>
-  <include location="resources.qrc"/>
+  <include location="../resources/icons.qrc"/>
  </resources>
  <connections/>
 </ui>
