--- Fotosopic/resource.qrc
@@ -1,12 +1,0 @@
-<RCC>
-    <qresource prefix="/images">
-        <file>delete.png</file>
-        <file>open.png</file>
-        <file>redo.png</file>
-        <file>save.png</file>
-        <file>undo.png</file>
-    </qresource>
-    <qresource prefix="/">
-        <file>../Downloads/icons/exit.png</file>
-    </qresource>
-</RCC>
