--- Fotosopic/resources.qrc
@@ -1,17 +1,0 @@
-<RCC>
-    <qresource prefix="/images">
-        <file>crop.jpeg</file>
-        <file>delete.png</file>
-        <file>draw.png</file>
-        <file>exit.png</file>
-        <file>mirror.png</file>
-        <file>open.png</file>
-        <file>redo.png</file>
-        <file>resize.png</file>
-        <file>save.png</file>
-        <file>text.png</file>
-        <file>undo.png</file>
-        <file>zoom.png</file>
-        <file>zoom_out.png</file>
-    </qresource>
-</RCC>
