+++ Fotosopic/resources/icons/delete.png
