+++ Fotosopic/resources/icons/draw.png
