+++ Fotosopic/resources/icons/redo.png
