+++ Fotosopic/resources/icons/resize.png
