+++ Fotosopic/resources/icons/undo.png
