+++ Fotosopic/resources/icons/zoom_out.png
